% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_summary.R
\name{cohort_summary}
\alias{cohort_summary}
\title{Summarise cohorts.}
\usage{
cohort_summary(id, mother, father, cohort = NULL, genotype = NULL,
  genotype_delim = "")
}
\arguments{
\item{id}{vector. Individual IDs}

\item{mother}{vector. Maternal IDs corresponding to id.}

\item{father}{vector. Paternal IDs corresponding to id.}

\item{cohort}{vector (optional). Cohort number (e.g. birth year) corresponding to the id.}

\item{genotype}{vector. Genotypes IDs corresponding to id.}

\item{genotype_delim}{char. A character denoting the genotype delimited. Default = "".}
}
\description{
Summarise cohorts by number of founders, proportion of individuals genotyped and/or allele frequency information at a given locus.
}
