% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genedrop.R
\name{genedrop}
\alias{genedrop}
\title{Conduct a single Gene-Drop Simulation.}
\usage{
genedrop(id, mother, father, cohort, genotype, nsim, n_founder_cohorts,
  fix_founders = T, verbose = T, return_genotypes = T,
  keep_only_genotyped_ids = T, delim = "")
}
\arguments{
\item{id}{vector. Individual IDs}

\item{mother}{vector. Maternal IDs corresponding to id. Missing mothers
should be coded as 0 or NA}

\item{father}{vector. Paternal IDs corresponding to id. Missing fathers
should be coded as 0 or NA}

\item{cohort}{vector. Cohort IDs corresponding to id.}

\item{genotype}{vector. Genotypes IDs corresponding to id.}

\item{nsim}{integer. Number of genedrop simulations to run.}

\item{n_founder_cohorts}{integer. The number of cohorts at the top of the
pedigree that will sample from the true allele frequences (these are
defined as "sampled"). All cohorts following these ones are "simulated" and
are used for comparisons of changes in allele frequency.}

\item{fix_founders}{logical. Default = TRUE. Determines whether individuals in
founder cohorts should be given their true recorded genotypes (if known).
For individuals with no known genotype, their genotypes are sampled based
on the observed cohort allele frequency. If FALSE, then all IDs are sampled
based on the cohort allele frequencies.}

\item{verbose}{logical. Default = TRUE. Output progress of the function.}

\item{return_genotypes}{logical. Default = TRUE. Output simulated genotypes for
each individual. If F, only returns the allele frequencies.}

\item{keep_only_genotyped_ids}{logical. Default = TRUE. Keeps only genotyped IDs
to allow a direct comparison of changes in allele frequencies.}
}
\description{
Conduct a single Gene-Drop Simulation.
}
