% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genedrop_freq.R
\name{genedrop_freq}
\alias{genedrop_freq}
\title{Conduct a single Gene-Drop Simulation with founder allele frequencies.}
\usage{
genedrop_freq(
  id,
  mother,
  father,
  cohort = NULL,
  allele_ids,
  founder_allele_freqs,
  nsim,
  genotype_delim = "",
  verbose = T,
  interval = 100
)
}
\arguments{
\item{id}{vector. Individual IDs}

\item{mother}{vector. Maternal IDs corresponding to id.}

\item{father}{vector. Paternal IDs corresponding to id.}

\item{cohort}{vector (optional). Cohort number (e.g. birth year)
corresponding to the id.}

\item{nsim}{integer. Number of genedrop simulations to run.}

\item{genotype_delim}{char. A character denoting the genotype delimited.
Default = "".}

\item{verbose}{logical. Output the progress of the run.}

\item{interval}{int. Default 100. Output progress every 100 simulations.}
}
\description{
Conduct a simple gene-drop analysis at a single locus, defining alleles and
allele frequencies in the founder cohort.
}
