% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genedrop_multi_warbler.R
\name{genedrop_multi_warbler}
\alias{genedrop_multi_warbler}
\title{Conduct a single Gene-Drop Simulation for a multi-allelic locus.}
\usage{
genedrop_multi_warbler(
  id,
  mother,
  father,
  cohort = NULL,
  genotype,
  genotype_delim = "",
  nsim,
  n_founder_cohorts = 1,
  fix_founders = T,
  verbose = T,
  interval = 100
)
}
\arguments{
\item{id}{vector. Individual IDs}

\item{mother}{vector. Maternal IDs corresponding to id.}

\item{father}{vector. Paternal IDs corresponding to id.}

\item{cohort}{vector (optional). Cohort number (e.g. birth year) corresponding to the id.}

\item{genotype}{vector. Genotypes IDs corresponding to id.}

\item{genotype_delim}{char. A character denoting the genotype delimited. Default = "".}

\item{nsim}{integer. Number of genedrop simulations to run.}

\item{n_founder_cohorts}{integer. The number of cohorts at the top of the
pedigree that will sample from the true allele frequences (these are
defined as "sampled"). All cohorts following these ones are "simulated" and
are used for comparisons of changes in allele frequency.}

\item{fix_founders}{logical. Default = TRUE. Determines whether individuals in
founder cohorts should be given their true recorded genotypes (if known).
For individuals with no known genotype, their genotypes are sampled based
on the observed cohort allele frequency. If FALSE, then all IDs are sampled
based on the cohort allele frequencies.}

\item{verbose}{logical. Output the progress of the run.}

\item{interval}{int. Default 100. Output progress every 100 simulations.}
}
\description{
Conduct a single Gene-Drop Simulation for a multi-allelic locus.
}
