% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genedrop_summary.R
\name{genedrop_summary}
\alias{genedrop_summary}
\title{Summarise allele frequencies from a genedrop object.}
\usage{
genedrop_summary(genedrop_object, genotype.delim = "")
}
\arguments{
\item{genedrop_object}{Output from genedrop_snp() or genedrop_multi()}

\item{genotype.delim}{char. Default = "". Delimiter character for genotypes.}
}
\description{
Summarise observed and simulated allele frequencies by simulation and cohort for a genedrop object.
}
