% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locus_summary.R
\name{locus_summary}
\alias{locus_summary}
\title{Summarise genotype frequencies by cohort in a pedigree at a single locus.}
\usage{
locus_summary(id, mother, father, cohort, genotype, na.include = F)
}
\arguments{
\item{id}{vector. Individual IDs}

\item{mother}{vector. Maternal IDs corresponding to id. Missing mothers
should be coded as 0 or NA}

\item{father}{vector. Paternal IDs corresponding to id. Missing fathers
should be coded as 0 or NA}

\item{cohort}{vector. Cohort IDs corresponding to id.}

\item{genotype}{vector. Genotypes IDs corresponding to id.}

\item{na.include}{boolean. Default FALSE. Indicate if NA cohorts are to be
output in the locus summary.}
}
\description{
Summarise genotype frequencies by cohort in a pedigree at a single locus.
}
