% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genedrop_plot_functions.R
\name{plot.genedroppeR}
\alias{plot.genedroppeR}
\title{Plot a summary graph of Gene-Drop Simulations.}
\usage{
\method{plot}{genedroppeR}(
  genedrop_obj,
  analysis_to_plot = "all",
  sim_alpha = 0.3,
  obs_line_col = "red"
)
}
\arguments{
\item{genedrop_obj}{Gene-Drop summary object from the function
`summary_genedrop()`}

\item{analysis_to_plot}{Default = "all". Options are "frequency",
"directional" and "cumulative". This will print graphs of allele frequency
changes, directional change, and cumulative change. Press <Enter> to
advance through each option. Alternatively, the user can specify options to show
specific graphs (e.g. "directional").}

\item{sim_alpha}{Default = 0.3. Alpha (transparency) value for plotted lines.}

\item{obs_line_col}{Default = "red". Line colour to use for the observed
data.}
}
\description{
Plot a summary graph of Gene-Drop Simulations.
}
