% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genedrop_plot_functions.R
\name{plot_genedrop}
\alias{plot_genedrop}
\title{`plot_genedrop()`: Plot a summary graph of `genedrop_...()` outputs.}
\usage{
plot_genedrop(
  genedrop_obj,
  analysis_to_plot = "all",
  sim_alpha = 0.3,
  obs_line_col = "red"
)
}
\arguments{
\item{genedrop_obj}{Gene-Drop object}

\item{analysis_to_plot}{Default = "all". Options are "frequency",
"directional" and "cumulative". This will output graphs of allele frequency
changes, directional change, and cumulative change. If the locus is
multi-allelic and the user specifies the default option ("all"), the user
will be prompted to press <Enter> to advance through each option.}

\item{sim_alpha}{Default = 0.3. Alpha (transparency) value for plotted lines.}

\item{obs_line_col}{Default = "red". Line colour to use for the observed
data.}
}
\value{
plot of output from `genedrop_...()` functions
}
\description{
`plot_genedrop()`: Plot a summary graph of `genedrop_...()` outputs.
}
\examples{
data(unicorn)
genedrop_obj <- genedrop_snp(
  id = unicorn$id,
  mother = unicorn$mother,
  father = unicorn$father,
  cohort = unicorn$cohort,
  genotype = unicorn$Horns,
  nsim = 100,
  n_founder_cohorts = 4,
  fix_founders = TRUE,
  verbose = TRUE,
  interval = 10,
  resample_offspring = FALSE
)
plot_genedrop(genedrop_obj)
}
