% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genedrop_plot_functions.R
\name{plot_genedrop_cohort}
\alias{plot_genedrop_cohort}
\title{`plot_genedrop_cohort()`: Plot a summary graph of Gene-Drop Simulations.}
\usage{
plot_genedrop_cohort(cohort_obj)
}
\arguments{
\item{cohort_obj}{data.frame created by `summary_cohort()`}
}
\description{
`plot_genedrop_cohort()`: Plot a summary graph of Gene-Drop Simulations.
}
