% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genedrop_plot_functions.R
\name{plot_genedrop_cumulative_change}
\alias{plot_genedrop_cumulative_change}
\title{Plot a histogram of Gene-Drop Simulation cumulative change values.}
\usage{
plot_genedrop_cumulative_change(
  genedrop_object_summary,
  n_founder_cohorts = NULL,
  remove_founders = T,
  obs_line_col = "red"
)
}
\arguments{
\item{genedrop_object_summary}{Gene-Drop summary object from the function
`summary_genedrop()`}

\item{n_founder_cohorts}{integer. The number of cohorts at the top of the
pedigree that will sample from the true allele frequences (these are
defined as "sampled"). All cohorts following these ones are "simulated" and
are used for comparisons of changes in allele frequency.}

\item{remove_founders}{logical. Default = TRUE. Remove the founders from the
data before calculating the linear regression slopes.}

\item{obs_line_col}{line colour to use for the observed data.}

\item{method}{Default = "lm". Currently the only option.}
}
\description{
Plot a histogram of Gene-Drop Simulation cumulated change values and return
true distribution values. This is the cumulative change in allele frequencies
over time and may be a signature of balancing selection.
}
