% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genedrop_plot_functions.R
\name{plot_genedrop_results}
\alias{plot_genedrop_results}
\title{Plot a summary graph of Gene-Drop Simulations.}
\usage{
plot_genedrop_results(
  genedrop_object_summary,
  sim_alpha = 0.2,
  obs_line_col = "red"
)
}
\arguments{
\item{genedrop_object_summary}{Gene-Drop summary object from the function `summary_genedrop()`}

\item{sim_alpha}{alpha (transparency) value for plotted points}

\item{obs_line_col}{line colour to use for the observed data.}
}
\description{
Plot a summary graph of Gene-Drop Simulations.
}
