% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genedrop_slopes.R
\name{plot_genedrop_slopes}
\alias{plot_genedrop_slopes}
\title{Plot Genedrop slopes.}
\usage{
plot_genedrop_slopes(genedrop_object)
}
\arguments{
\item{genedrop_object}{Output from genedrop() function}
}
\description{
Plot slopes from a genedrop analysis
}
