% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_locus_summary.R
\name{plot_locus_summary}
\alias{plot_locus_summary}
\title{Plot locus summary output}
\usage{
plot_locus_summary(x)
}
\arguments{
\item{x}{Output of locus_summary() function}
}
\description{
Plot locus summary output to investigate proportion of IDs genotyped,
proportion of founder individuals and the total ID count per cohort.
}
