% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_genedrop.R
\name{summarise_genedrop}
\alias{summarise_genedrop}
\title{Summarise allele frequencies from a genedrop object.}
\usage{
summarise_genedrop(genedrop_obj, genotype_delim = "")
}
\arguments{
\item{genedrop_obj}{An object of class "genedroppeR" as output by functions
`genedrop_...()`.}

\item{genotype_delim}{char. Default = ''. Delimiter character for genotypes.}
}
\description{
Summarise observed and simulated allele frequencies by simulation and cohort
for a genedrop object. This is only called within specific functions and is
not for general use.
}
