% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_genedrop.R
\name{summary.genedroppeR_cohort}
\alias{summary.genedroppeR_cohort}
\title{Generic summary function for class "genedroppeR_cohort"}
\usage{
\method{summary}{genedroppeR_cohort}(cohort_obj, ...)
}
\arguments{
\item{cohort_obj}{An object of class "cohort_obj" as output by
function `summary_cohort()`.}

\item{...}{Further arguments to be passed}
}
\description{
Generic summary function for class "genedroppeR_cohort"
}
