% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_cohort.R
\name{summary_cohort}
\alias{summary_cohort}
\title{Summarise cohorts.}
\usage{
summary_cohort(
  id,
  mother = NULL,
  father = NULL,
  cohort = NULL,
  genotype = NULL,
  genotype_delim = ""
)
}
\arguments{
\item{id}{vector. Individual IDs}

\item{mother}{vector. Maternal IDs corresponding to id.}

\item{father}{vector. Paternal IDs corresponding to id.}

\item{cohort}{vector (optional). Cohort number (e.g. birth year)
corresponding to the id.}

\item{genotype}{vector. Genotypes IDs corresponding to id.}

\item{genotype_delim}{char. A character denoting the genotype delimited.
Default = ''.}
}
\description{
Summarise cohorts by number of founders, proportion of individuals genotyped
and/or allele frequency information at a given locus.
}
\examples{

data(unicorn)
x <- summary_cohort(id = unicorn$id,
                    mother = unicorn$mother,
                    father = unicorn$father,
                    cohort = unicorn$cohort,
                    genotype = unicorn$Horns)

}
