% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unicorn.R
\docType{data}
\name{unicorn}
\alias{unicorn}
\title{Example dataset for genedroppeR examples and tutorial}
\format{
## `unicorn`
\describe{
  \item{id}{Individual IDs}
  \item{mother}{Mother IDs}
  \item{father}{Father IDs}
  \item{cohort}{Cohort identifiers (e.g. year of birth)}
  \item{sex}{Sex (1 = heterogametic sex, 2 = homogametic sex)}
  \item{Horns}{A biallelic locus for horn length}
  \item{MHC}{A multi-allelic locus for the magic histocompatibility locus}
  \item{Glitter}{A biallelic locus underpinning a rare glitter coat polymorphism}
  \item{Xlinked}{A biallelic, sex-linked locus.}
}
}
\usage{
unicorn
}
\description{
This dataset is from a long-term study of unicorns on the island of Áiteigin
in Scotland. They have been genotyped for four loci: HornSNP, a biallelic
locus for horn length; MHC, which characterises multi-allelic variation at
the Magic Histocompatibility Locus; Glitter, a SNP responsible for a rare
glitter coat polymorphism; and Xlinked, an X chromosome locus.
}
\keyword{dataset}
