% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_data}
\alias{check_data}
\title{Check data before gene-drop analysis.}
\usage{
check_data(id, mother, father, cohort = NULL, genotype)
}
\arguments{
\item{id}{vector. Individual IDs}

\item{mother}{vector. Maternal IDs corresponding to id.}

\item{father}{vector. Paternal IDs corresponding to id.}

\item{cohort}{vector (optional). Cohort number (e.g. birth year)
corresponding to the id.}

\item{genotype}{vector. Genotypes corresponding to id.}
}
\description{
Check data before gene-drop analysis.
}
