% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genedrop_snp.R
\name{genedrop_snp}
\alias{genedrop_snp}
\title{Conduct a single Gene-Drop Simulation for a biallelic locus.}
\usage{
genedrop_snp(
  id,
  mother,
  father,
  cohort = NULL,
  genotype,
  nsim,
  n_founder_cohorts = 1,
  fix_founders = T,
  verbose = T,
  interval = 100,
  resample_offspring = F
)
}
\arguments{
\item{id}{vector. Individual IDs}

\item{mother}{vector. Maternal IDs corresponding to id.}

\item{father}{vector. Paternal IDs corresponding to id.}

\item{cohort}{vector (optional). Cohort number (e.g. birth year)
corresponding to the id.}

\item{genotype}{vector. Genotypes corresponding to id.}

\item{nsim}{integer. Number of genedrop simulations to run.}

\item{n_founder_cohorts}{integer. The number of cohorts at the top of the
pedigree that will sample from the true allele frequencies (these are
defined as "sampled"). All cohorts following these ones are "simulated" and
are used for comparisons of changes in allele frequency.}

\item{fix_founders}{logical. Default = TRUE. Determines whether individuals
in founder cohorts should be given their true recorded genotypes (if
known). For individuals with no known genotype, their genotypes are sampled
based on the observed cohort allele frequency. If FALSE, then all IDs are
sampled based on the cohort allele frequencies.}

\item{verbose}{logical. Default = TRUE. Output the progress of the run.}

\item{interval}{integer. Default 100. Output progress every 100 simulations.}

\item{resample_offspring}{logical. Default = FALSE. If FALSE, the same
pedigree structure as the observed pedigree is used. If TRUE, then
offspring are resampled across parents in each cohort. This is to remove
any potential signal where prolific individuals tend to have prolific
offspring, but will also mean that pedigrees are not directly comparable.}
}
\description{
Conduct a single Gene-Drop Simulation for a biallelic locus.
}
