% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genedroppeR-package.R
\docType{package}
\name{genedroppeR-package}
\alias{genedroppeR}
\alias{genedroppeR-package}
\title{genedroppeR: genedroppeR: conduct single-locus genedrop analyses through pedigrees}
\description{
This package conducts gene-drop analyses through pedigrees.
}
