% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_genedrop.R
\name{summary.genedroppeR}
\alias{summary.genedroppeR}
\title{Generic summary functions for class "genedroppeR" and "genedroppeR_cohort"}
\usage{
\method{summary}{genedroppeR}(genedrop_obj)
}
\description{
Generic summary functions for class "genedroppeR" and "genedroppeR_cohort"
}
