% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_genedrop.R
\name{summary_genedrop}
\alias{summary_genedrop}
\title{Generic summary function for class "genedroppeR"}
\usage{
summary_genedrop(genedrop_obj)
}
\arguments{
\item{genedrop_obj}{An object of class "genedroppeR" as output by functions `genedrop_...()`.}
}
\description{
Generic summary function for class "genedroppeR"
}
