% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unicorn.R
\docType{data}
\name{unicorn}
\alias{unicorn}
\title{Example dataset for genedroppeR examples and tutorial}
\format{
## `unicorn`
\describe{
  \item{id}{Individual IDs}
  \item{father}{Father IDs}
  \item{mother}{Mother IDs}
  \item{cohort}{Cohort identifiers (e.g. year of birth)}
  \item{cohort}{Sex (1 = heterogametic sex, 2 = homogametic sex)}
  \item{HornSNP}{A biallelic locus for horn length}
  \item{MHC}{A multi-allelic locus for the magic histocompatibility locus}
  \item{ColourSNP}{A biallelic locus underpinning a rare glitter coat polymorphism}
  \item{SexLinkedGeno}{A biallelic, sex-linked locus.}
}
}
\usage{
unicorn
}
\description{
This dataset is from a long-term study of unicorns on the
island of Áiteigin in Scotland. They have been genotyped for three
loci: HornSNP, a biallelic locus for horn length; MHC, which characterises
multi-allelic variation at the Magic Histocompatibility Locus; and ColourSNP,
a SNP responsible for a rare glitter coat polymorphism.
}
\keyword{dataset}
